refactor: move create label to separate component and add enter key create support
