+++ application/ui/src/features/project/label-selection/validator.ts
@@ -5,10 +5,6 @@ import { Label } from '../../../constants/shared-types';
 import { convertHotkeyToOSFormat } from '../../../shared/hotkeys-definition';
 
 export const validateLabelName = (label: Label, labels: Label[]): string | undefined => {
-    if (label.name.trim().length === 0) {
-        return 'Label name cannot be empty';
-    }
-
     if (labels.some((l) => l.name === label.name)) {
         return 'That label name already exists';
     }
@@ -18,8 +14,6 @@ export const validateLabelName = (label: Label, labels: Label[]): string | undef
 
 export const validateLabelHotkey = (hotkey: string, allHotkeys: string[]) => {
     const osFormatHotkeys = allHotkeys.map(convertHotkeyToOSFormat).map((key) => key.toLowerCase());
-
-    console.log({ osFormatHotkeys, hotkey });
 
     if (osFormatHotkeys.includes(hotkey.toLowerCase())) {
         return 'That hotkey is already in use';
