+++ application/ui/src/routes/project/create-project.tsx
@@ -4,12 +4,12 @@
 import { View } from '@geti/ui';
 
 import { CreateProjectForm } from '../../features/project/create/create-project-form';
-import { useProjectsQuery } from '../../hooks/api/project.hook';
+import { useProjects } from '../../hooks/api/project.hook';
 
 import backgroundStyles from '../../features/project/project-background.module.scss';
 
 export const CreateProject = () => {
-    const { data: projects = [] } = useProjectsQuery();
+    const { data: projects = [] } = useProjects();
 
     return (
         <View UNSAFE_className={backgroundStyles.projectBackground} height='100%' width='100%'>
